// src/app/shared/interfaces/historico-cobranca.model.ts
export interface HistoricoCobrancaMV {
  style: string;
  clienteId: number;
  nomeFantasia: string;
  placa: string;
  dataHoraAtendimento: number | string;
  dataHoraFechamento: number | string;
  dataEnvioCobranca: number | string;
  dataResposta: number | string;
  status: string;
  atendimentoId: number;
  veiculoModelo: string;
  totalPagina: 0;
}